+++ .idea/workspace.xml
@@ -2,11 +2,7 @@
 <project version="4">
   <component name="ChangeListManager">
     <list default="true" id="4cd1f0cf-e790-4da8-8e26-2da16a212d1e" name="Default Changelist" comment="">
-      <change beforePath="$PROJECT_DIR$/admin/reset_info.php" beforeDir="false" afterPath="$PROJECT_DIR$/admin/reset_info.php" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/data_creation.sql" beforeDir="false" afterPath="$PROJECT_DIR$/data_creation.sql" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/func/db.php" beforeDir="false" afterPath="$PROJECT_DIR$/func/db.php" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/func/func.php" beforeDir="false" afterPath="$PROJECT_DIR$/func/func.php" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/static/functions.js" beforeDir="false" afterPath="$PROJECT_DIR$/static/functions.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -29,6 +25,7 @@
     <property name="WebServerToolWindowFactoryState" value="false" />
     <property name="last_opened_file_path" value="$PROJECT_DIR$" />
     <property name="nodejs_package_manager_path" value="npm" />
+    <property name="settings.editor.selected.configurable" value="preferences.lookFeel" />
     <property name="vue.rearranger.settings.migration" value="true" />
   </component>
   <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
@@ -39,11 +36,17 @@
       <option name="number" value="Default" />
       <option name="presentableId" value="Default" />
       <updated>1607626663857</updated>
-      <workItem from="1607626664949" duration="2422000" />
+      <workItem from="1607626664949" duration="3570000" />
     </task>
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
     <option name="version" value="3" />
+  </component>
+  <component name="WindowStateProjectService">
+    <state x="279" y="58" key="SettingsEditor" timestamp="1607630753511">
+      <screen x="0" y="0" width="1536" height="824" />
+    </state>
+    <state x="279" y="58" key="SettingsEditor/0.0.1536.824@0.0.1536.824" timestamp="1607630753511" />
   </component>
 </project>
