+++ func/dbQuery.php
